% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_custom_style.R
\name{add_custom_style}
\alias{add_custom_style}
\title{Include Custom HTML Style Elements Into Active Document}
\usage{
add_custom_style(main = TRUE, ...)
}
\arguments{
\item{main}{Logical. Include \emph{summarytools.css} file. \code{TRUE} by
default. This will affects only \strong{summarytools} objects, for one
exception: two properties of the \code{img} tag are redefined to have
\code{background-color: transparent} and \code{border: 0}.}

\item{\dots}{Character. Path to additional \emph{css} file(s) to include.}
}
\value{
The style content and \emph{css} file(s) returned silently as a
character vector, and printed (using \code{cat()}).
}
\description{
Generate inline style components to modify the standard flexdashboard
\emph{css}. Original functionality updated based on example from
\code{st_css()} from \url{https://github.com/dcomtois/summarytools}
}
\details{
The style formatting is intended to be used with bootsrap's
\code{readable} theme.

The function may be called in an \emph{R markdown} document code
chunk having options \code{echo=FALSE} and \code{results="asis"}.
}
