% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjective_animal.R
\name{adjective_animal}
\alias{adjective_animal}
\title{Return a name comprised of an adjective animal word pair}
\arguments{
\item{n}{number of word pairs to return}
}
\value{
vector of character values containting adjective_animal names
}
\description{
Return a name comprised of an adjective animal word pair
}
\examples{
adjective_animal(3)
}
