% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjectives.R
\docType{data}
\name{adjectives}
\alias{adjectives}
\title{Adjectives.}
\format{A vector containing 214 adjectives}
\source{
<https://github.com/rstudio/cranwhales>
}
\usage{
adjectives
}
\description{
Select English language adjectives.
}
\keyword{datasets}
