% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignx_n.R
\name{alignx_n}
\alias{alignx_n}
\title{Right align numeric values}
\usage{
alignx_n(x, num_dec = 3, pad_char = "&numsp;", do_roundx_n = TRUE)
}
\arguments{
\item{x}{numeric value to be right justified}

\item{num_dec}{number of decimal places to justify}

\item{pad_char}{padding character defaults to \code{&numsp;}}

\item{do_roundx_n}{logical to determine if \code{roundx_n} is performed.
Defaults to True}
}
\value{
character
}
\description{
Evaluate a numeric value and optionally perform \code{roundx_n} before
padding the decimal portion of the number to n digits. Note: this coverts the
numeric value to a character. The function is intended only for final display
purposes.
}
\examples{
cat(alignx_n(c(0.12345, 1.123344, 100.12345)), sep = "\n")

# Without \code{roundx_n}
cat(alignx_n(c(100.123, 100.23, 100.1), do_roundx_n = FALSE), sep = "\n")
}
