% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animals.R
\docType{data}
\name{animals}
\alias{animals}
\title{Animals.}
\format{A vector containing 178 animals}
\source{
<https://github.com/rstudio/cranwhales>
}
\usage{
animals
}
\description{
Select English language animals.
}
\keyword{datasets}
