% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_toc_entries.R
\name{create_toc_entries}
\alias{create_toc_entries}
\title{Create table of contents entries from web page content}
\usage{
create_toc_entries(html_file)
}
\arguments{
\item{html_file}{a html file}
}
\value{
A character vector of table of formatted table contents entries
}
\description{
Create table of contents entries from web page content
}
