% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curios-package.R
\docType{package}
\name{curios-package}
\alias{curios}
\alias{curios-package}
\title{curios: Curios Used in Support of Practical Data Science}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

A compendium of functions and data to support hands-on data science particularly. This is not a formal R package, but rather a collection of tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/johnaclouse/curios}
  \item Report bugs at \url{https://github.com/johnaclouse/curios/issues}
}

}
\author{
\strong{Maintainer}: John Clouse \email{john@clouse.io} (\href{https://orcid.org/0000-0002-0369-2116}{ORCID})

}
\keyword{internal}
