% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_case.R
\name{fix_case}
\alias{fix_case}
\title{Fix word case}
\usage{
fix_case(
  x,
  to_lower_first = FALSE,
  replace_underscores = TRUE,
  underscore_replacement = " ",
  to_lower_minor_words = TRUE,
  minor_words = curios::minor_words,
  to_upper_known_acronyms = TRUE,
  known_acronyms = curios::known_acronyms,
  replace_named_words = FALSE,
  named_words = NA,
  case = c("title", "sentence", "lower", "upper")
)
}
\arguments{
\item{x}{character string.}

\item{to_lower_first}{logical,indicating character string should first be converted to all lower case.}

\item{replace_underscores}{logical determines replacement of underscore characters.}

\item{underscore_replacement}{character string to replace underscores.}

\item{to_lower_minor_words}{logical determines if minor words are all lower case.}

\item{minor_words}{named character vector of minor words that will be converted to lower case.}

\item{to_upper_known_acronyms}{logical determines if known acronyms are replaced with case correct versions.}

\item{known_acronyms}{named character vector of acronyms that will be converted to proper case.}

\item{replace_named_words}{logical determines if a set of \code{named_words} are replaced with case correct versions.}

\item{named_words}{named character vector of words that will be replaced with alternate words or word case}

\item{case}{character designating the overall case structure: title (default), sentence, lower, upper.}
}
\value{
character string
}
\description{
Attempt to convert words from given format to sentence, title, lower, or
upper case while being mindful of APA style guidelines, acronyms, and custom
word replacement lists.
}
\details{
Note: word replacement is done in a case sensitive manner.  For
example, "usDA" would be replaced with "USDA" as would "UsDA" or "usda".
}
\examples{
fix_case("the hba1c level was reported as a1C", case = "sentence")
}
