% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_toc_entries.R
\name{format_toc_entry}
\alias{format_toc_entry}
\title{Format individual table of contents line}
\usage{
format_toc_entry(x, html_file)
}
\arguments{
\item{x}{a HTML node}

\item{html_file}{URL for the page containing the node}
}
\value{
a character value with the table of contents entry
}
\description{
Format individual table of contents line
}
\keyword{internal}
