% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_discrete.R
\name{is_discrete}
\alias{is_discrete}
\title{Check to see if a numeric vector is discrete based number of unique values}
\usage{
is_discrete(x, cutoff = 100)
}
\arguments{
\item{x}{a vector of numeric values}

\item{cutoff}{minimum number of uniqu}
}
\value{
logical
}
\description{
A simple test intended to distinguish between discrete and continuous values
for the purpose of choosing analytics visualization selection. The selection
of a cutoff value is purely subjective.
}
\examples{
is_discrete(1:10)
is_discrete(runif(1000))
}
