% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{known_acronyms}
\alias{known_acronyms}
\title{Known acronyms}
\format{
A named character vector
}
\usage{
known_acronyms
}
\description{
A named character vector for use with stringr::str_replace_all to replace
lower cased versions of acronyms with an all capitalized version of the
acronym.
}
\keyword{datasets}
