% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{minor_words}
\alias{minor_words}
\title{Minor words}
\format{
A character vector
}
\source{
\url{https://apastyle.apa.org/style-grammar-guidelines/capitalization/title-case}
}
\usage{
minor_words
}
\description{
A named character vector containing minor words as defined in the \href{https://apastyle.apa.org/style-grammar-guidelines/capitalization/title-case}{APA style guide}.
The words in this set are not to be capitalized in title case.
}
\keyword{datasets}
