% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_all_code_chunks.R
\name{remove_code_chunk_names}
\alias{remove_code_chunk_names}
\title{Remove generic code chunk names from an RMD file}
\usage{
remove_code_chunk_names(x)
}
\arguments{
\item{x}{a complete path, filename, and extension to a .Rmd file}
}
\description{
This function is a wrapper for /code{namer::unname_all_chunks}
}
