% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_all_code_chunks.R
\name{rename_all_code_chunks}
\alias{rename_all_code_chunks}
\title{Rename all generic code chunks in current directory}
\usage{
rename_all_code_chunks()
}
\description{
This function changes source code files! All generic code chunks created
using the file name as a prefix will be removed and new code chunks names
created using /code{namer::name_chunks}. Code chunks with specific names
should not be effected.
}
\examples{
# rename_all_code_chunks()
}
