% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundx_n.R
\name{roundx_n}
\alias{roundx_n}
\title{Round value based on magnitude}
\usage{
roundx_n(x)
}
\arguments{
\item{x}{A numeric value to be rounded. Values >= 100 are rounded to 1
decimal place. Values < 100 and >=1 are rounded to two decimal places.
Values <1 are rounded to three decimal places.}
}
\value{
A numeric value rounded based on the magnitude of \code{x}
}
\description{
Round value based on magnitude
}
\examples{
library(curios)
# Note: R will display right padded zero's below. This is a display manifestation only.
roundx_n(c(0.1234, 1.1234, 100.123))
# Verify rounding does not add right hand padding
roundx_n(c(0.1234, 1.1234, 100.123))[3]
}
