% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentance_case.R
\name{sentance_case}
\alias{sentance_case}
\title{sentance_case}
\usage{
sentance_case(x)
}
\arguments{
\item{x}{character}
}
\value{
character
}
\description{
sentance_case
}
\examples{
sentance_case("the CAT's name is Skipper")
}
