% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_toc_entry.R
\name{format_toc_entry}
\alias{format_toc_entry}
\title{Format individual table of contents line}
\usage{
format_toc_entry(entry, padding = "2em", html_file = "")
}
\arguments{
\item{entry}{a character string to be converted to a Table of Contents entry.}

\item{padding}{a character string to be added to the front of the entry string
typically for indentation purposes.}

\item{html_file}{URL for the page containing the node. Defaults to \code{""}
indicating current page.}
}
\description{
The function is used for its side-effect of printing a table of contents
entry.
}
\examples{
format_toc_entry(c("Header 1", "Header 2"))
format_toc_entry("Header 1", "---->", "mypage.html")
}
\keyword{internal}
