% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_toc_entry.R
\name{format_toc_entry}
\alias{format_toc_entry}
\title{Format individual table of contents line}
\usage{
format_toc_entry(entry, prefix = "|    ", html_file = "")
}
\arguments{
\item{entry}{a character string to be converted to a Table of Contents entry.}

\item{prefix}{a character string to be added to the front of the entry string
typically for indentation purposes.}

\item{html_file}{URL for the page containing the node. Defaults to \code{""}
indicating current page.}
}
\value{
a character value with the table of contents entry
}
\description{
Format individual table of contents line
}
\keyword{internal}
